jsc.bless documentation and example
