+++ lib/arbitraryBless.js
@@ -20,10 +20,6 @@ function arbitraryProtoSMap(f, g, newShow) {
 /* eslint-enable no-use-before-define */
 
 /**
-  - `arb.bless({...}): arbitrary a`
-
-      Bless generator, shrink, show triple with  with `.smap` property.
-
   - `.smap(f: a -> b, g: b -> a, newShow: (b -> string)?): arbitrary b`
 
       Transform `arbitrary a` into `arbitrary b`. For example:
