+++ lib/bless.js
@@ -1,0 +1,47 @@
+"use strict";
+
+var assert = require("assert");
+
+var arbitraryBless = require("./arbitraryBless.js");
+var generator = require("./generator.js");
+var show = require("./show.js");
+var shrink = require("./shrink.js");
+
+/**
+  - `bless(arb: {...}): arbitrary a`
+
+    Bless almost arbitrary structure to be proper arbitrary. *Note*: this function mutates argument.
+
+    Example:
+
+    ```js
+    var arbTokens = jsc.bless({
+      generator: function () {
+        switch (jsc.random(0, 2)) {
+          case 0: return "foo";
+          case 1: return "bar";
+          case 2: return "quux";
+        }
+      }
+      shrink: jsc.shrink.noop,
+    });
+    ```
+*/
+function bless(arb) {
+  assert(arb !== null && typeof arb === "object", "bless: arb should be an object");
+  assert(typeof arb.generator === "function", "bless: arb.generator should be a function");
+  assert(typeof arb.shrink === "function", "bless: arb.shrink should be a function");
+
+  generator.bless(arb.generator);
+  shrink.bless(arb.shrink);
+
+  // default show
+  if (typeof arb.show !== "function") {
+    arb.show = show.def;
+  }
+
+  arbitraryBless(arb);
+  return arb;
+}
+
+module.exports = bless;
