+++ lib/jsverify.js
@@ -95,7 +95,7 @@ var assert = require("assert");
 var lazyseq = require("lazy-seq");
 
 var arbitrary = require("./arbitrary.js");
-var arbitraryBless = require("./arbitraryBless.js");
+var bless = require("./bless.js");
 var either = require("./either.js");
 var environment = require("./environment.js");
 var FMap = require("./finitemap.js");
@@ -364,28 +364,6 @@ function compile(str, env) {
 }
 
 /**
-  - `bless(arb: {...}): arbitrary a`
-
-    Bless almost arbitrary structure to be proper arbitrary. *Note*: this function mutates argument.
-*/
-function bless(arb) {
-  assert(arb !== null && typeof arb === "object", "bless: arb should be an object");
-  assert(typeof arb.generator === "function", "bless: arb.generator should be a function");
-  assert(typeof arb.shrink === "function", "bless: arb.shrink should be a function");
-
-  generator.bless(arb.generator);
-  shrink.bless(arb.shrink);
-
-  // default show
-  if (typeof arb.show !== "function") {
-    arb.show = show.def;
-  }
-
-  arbitraryBless(arb);
-  return arb;
-}
-
-/**
   - `sampler(arb: arbitrary a, genSize: nat = 10): (sampleSize: nat?) -> a`
 
       Create a sampler for a given arbitrary with an optional size. Handy when used in
@@ -433,6 +411,7 @@ function sampler(arb, size) {
 
 /// include ./typify.js
 /// include ./arbitraryBless.js
+/// include ./bless.js
 /// include ./primitive.js
 /// include ./arbitrary.js
 /// include ./fn.js
